# backend/openapi.yaml
openapi: "3.0.3"
info:
  version: 0.9.0
  title: Delightool server API
  contact:
    name: Delightool contact
    email: contact@delightool.dev
    url: https://delightool.dev

  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: /api
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /resources/create:
    post:
      summary: create resource
      operationId: createResource
      tags:
        - resource
      requestBody:
        description: new resource
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResource"
      responses:
        "201":
          description: resource created
          content:
            text/plain:
              schema:
                type: string
...
  /resources/type/create:
    post:
      summary: create resource_type
      operationId: createResourceType
      tags:
        - resource
      requestBody:
        description: new resource_type
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceType"
      responses:
        "201":
          description: resource_type created
          content:
            text/plain:
              schema:
                type: string
  /resources/type/delete/{path}:
    delete:
      summary: delete resource_type
      operationId: deleteResourceType
      tags:
        - resource
      parameters:
        - $ref: "#/components/parameters/Path"
      responses:
        "200":
          description: resource_type deleted
          content:
            text/plain:
              schema:
                type: string
  /resources/type/update/{path}:
    post:
      summary: update resource_type
      operationId: updateResourceType
      tags:
        - resource
      parameters:
        - $ref: "#/components/parameters/Path"
      requestBody:
        description: updated resource_type
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditResourceType"
      responses:
        "200":
          description: resource_type updated
          content:
            text/plain:
              schema:
                type: string
  /resources/type/get/{path}:
    get:
      summary: get resource_type
      operationId: getResourceType
      tags:
        - resource
      parameters:
        - $ref: "#/components/parameters/Path"
      responses:
        "200":
          description: resource_type deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceType"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  parameters:
...
    Path:
      name: path
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    CreateResource:
      type: object
      properties:
        path:
          type: string
        value:
          type: object
        description:
          type: string
        resource_type:
          type: string
      required:
        - path
        - value
        - resource_type
...
    ResourceType:
      type: object
      properties:
        name:
          type: string
        schema: {}
        description:
          type: string
      required:
        - name
    EditResourceType:
      type: object
      properties:
        schema:
          type: string
        description:
          type: string
...